refactor: Implement Mapper and migrate monetary fields to BigDecimal
refactor: migrate price and totals to BigDecimal and improve monetary calculations

- Replaced Double with BigDecimal in Sale and SaleDetail entities for proper monetary handling
- Updated DTOs to use BigDecimal consistently across the domain
- Improved subtotal calculation using BigDecimal.multiply() with proper scale and RoundingMode.HALF_UP
- Refactored total calculation to avoid floating-point inaccuracies
- Cleaned up mapper logic and removed unsafe/unnecessary imports
- Ensured DTO building uses precise arithmetic and scale normalization
- Minor naming fixes to match domain model conventions

+++ TechTestSupermarket/src/main/java/com/springmarkets/TechTestSupermarket/model/Sale.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -19,7 +20,7 @@ public class Sale {
   private Long id;
   private LocalDate date;
   private String status;
-  private Double total;
+  private BigDecimal total;
 
   //Relationship with the store
   @ManyToOne
