+++ TechTestSupermarket/src/main/java/com/springmarkets/TechTestSupermarket/model/SaleDetail.java
@@ -9,6 +9,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter @Setter
 @AllArgsConstructor
 @NoArgsConstructor
@@ -24,6 +26,6 @@ public class SaleDetail {
   //Product
   @ManyToOne
   private Product product;
-  private int prodQuantity;
-  private double unitPrice;
+  private Integer prodQuantity;
+  private BigDecimal unitPrice;
 }
